mcfg_format: fix invalid construction of embdes
When a dollar sign is encountered anywhere before an opening brace
is encountered, mcfg_format would think that is trying to build an
embed until an opening brace was encountered.

Fixed by setting building_embed to false as soon as any character
which is not an opening brace is encountered.